app.js screen component 별로 분리, homescreen 추가
